requirements.txt
Django==1.7.6
DjangoLightbox==0.1.2
Pillow==2.7.0
Unidecode==0.04.17
argparse==1.2.1
boto==2.38.0
diff-match-patch==20121119
django-audit-log==0.7.0
django-autoslug==1.7.2
django-bootstrap3==5.1.1
django-bootstrap3-datetimepicker==2.2.3
django-cities-light==3.1.0
django-configfield==0.2.2
django-contrib-comments==1.5
django-cuser==2014.9.28
django-datetime-widget==0.9.3
django-debug-toolbar==1.3.0
django-extensions==1.5.3
django-filer==0.9.9
django-form-admin==0.4.2
django-jquery==1.9.1
django-jquery-lightbox==2.6
django-jsonfield==0.8.11
django-localflavor-us==1.1
django-mailgun==0.2.2
django-mptt==0.6.1
django-multi-email-field==0.3.1
django-multiupload==0.3
django-polymorphic==0.6.1
django-relatedadminwidget==0.0.3
django-reversion==1.8.5
django-ses==0.7.0
django-suit==0.2.12
djangorestframework==3.1.1
easy-thumbnails==2.2
gunicorn==19.3.0
mock==1.0.1
openpyxl==2.2.1
psycopg2==2.6
python-dateutil==2.4.2
pytz==2014.10
redis==2.10.3
requests==2.7.0
setproctitle==1.1.8
six==1.9.0
sorl-thumbnail==12.2
sqlparse==0.1.11
wsgiref==0.1.2
xlrd==0.9.3
